Team permissions

+++ app/Http/Controllers/Api/TeamNoticeController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use App\Models\TeamNotice;
+use App\Models\{Team, TeamNotice};
 use Illuminate\Http\{Request, Response};
 use Illuminate\Support\Facades\Auth;
 
@@ -14,10 +14,19 @@ class TeamNoticeController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index($teamId)
+    public function index($team)
     {
+        /** @var \App\Models\User $user */
+        $user = Auth::guard('sanctum')->user();
+        $team = Team::find($team);
+
+        $user->hasTeamPermission($team, 'read');
+
+        $userCanRead = $user->hasTeamPermission($team, 'read');
+        abort_unless($userCanRead, Response::HTTP_FORBIDDEN);
+
         return TeamNotice::with(['author'])
-            ->where('team_id', $teamId)
+            ->where('team_id', $team->id)
             ->get();
     }
 
@@ -29,14 +38,21 @@ class TeamNoticeController extends Controller
      */
     public function store(Request $request, $team)
     {
+        /** @var \App\Models\User $user */
+        $user = Auth::guard('sanctum')->user();
+        $team = Team::find($team);
+
+        $userCanCreate = $user->hasTeamPermission($team, 'create');
+        abort_unless($userCanCreate, Response::HTTP_FORBIDDEN);
+
         $data = $request->validate([
             'title'     => ['required', 'string', 'max:255'],
             'body'      => ['required', 'string']
         ]);
 
         $notice = new TeamNotice($data);
-        $notice->author_id = Auth::guard('sanctum')->id();
-        $notice->team_id = $team;;
+        $notice->author_id = $user->id;
+        $notice->team_id = $team->id;
         $notice->save();
 
         return response()->json($notice->load(['author'])->makeHidden(['body']), Response::HTTP_CREATED);
@@ -46,15 +62,22 @@ class TeamNoticeController extends Controller
      * Display the specified resource.
      *
      * @param  int  $team
-     * @param  int  $id
+     * @param  int  $notice
      * @return \Illuminate\Http\Response
      */
-    public function show($team, $id)
+    public function show($team, $notice)
     {
+        /** @var \App\Models\User $user */
+        $user = Auth::guard('sanctum')->user();
+        $team = Team::find($team);
+
+        $userCanRead = $user->hasTeamPermission($team, 'read');
+        abort_unless($userCanRead, Response::HTTP_FORBIDDEN);
+
         return TeamNotice::with(['author'])
             ->where('author_id', Auth::guard('sanctum')->id())
-            ->where('team_id', $team)
-            ->findOrFail($id);
+            ->where('team_id', $team->id)
+            ->findOrFail($notice);
     }
 
     /**
@@ -62,11 +85,18 @@ class TeamNoticeController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $team
-     * @param  int  $id
+     * @param  int  $notice
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $team, $id)
+    public function update(Request $request, $team, $notice)
     {
+        /** @var \App\Models\User $user */
+        $user = Auth::guard('sanctum')->user();
+        $team = Team::find($team);
+
+        $userCanUpdate = $user->hasTeamPermission($team, 'update');
+        abort_unless($userCanUpdate, Response::HTTP_FORBIDDEN);
+
         $data = $request->validate([
             'title'     => ['nullable', 'string', 'max:255'],
             'body'      => ['nullable', 'string']
@@ -75,8 +105,8 @@ class TeamNoticeController extends Controller
         /** @var \App\Models\Notice $notice */
         $notice = TeamNotice::with(['author'])
             ->where('author_id', Auth::guard('sanctum')->id())
-            ->where('team_id', $team)
-            ->findOrFail($id);
+            ->where('team_id', $team->id)
+            ->findOrFail($notice);
 
         $notice->update([
             'title'     => $data['title'] ?? $notice->title,
@@ -90,16 +120,23 @@ class TeamNoticeController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $team
-     * @param  int  $id
+     * @param  int  $notice
      * @return \Illuminate\Http\Response
      */
-    public function destroy($team, $id)
+    public function destroy($team, $notice)
     {
+        /** @var \App\Models\User $user */
+        $user = Auth::guard('sanctum')->user();
+        $team = Team::find($team);
+
+        $userCanDelete = $user->hasTeamPermission($team, 'delete');
+        abort_unless($userCanDelete, Response::HTTP_FORBIDDEN);
+
         /** @var \Illuminate\Database\Eloquent\Model $notice */
         $notice = TeamNotice::with(['author'])
             ->where('author_id', Auth::guard('sanctum')->id())
-            ->where('team_id', $team)            
-            ->findOrFail($id);
+            ->where('team_id', $team->id)
+            ->findOrFail($notice);
 
         $notice->delete();
         return response()->json(null, Response::HTTP_NO_CONTENT);
